Use textures for compute storage and refine event management
Prior to this change, the compute pipeline were using buffers which are
not compatible with WebGPU.

This change subsitutes the buffers for textures. Also, additional
keyboard events are added for Ctrl+R and Ctrl+Q and a new mask task is
added that watches for changes, rebuilds the wasm files, starts the
serve task and opens a correctly configured chromium window.